+++ packages/react/src/components/OptionsMenu/OptionsMenuWrapper.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { OptionsMenuAlignmentProps } from './OptionsMenu';
 
@@ -29,11 +28,5 @@ const OptionsMenuWrapper = ({
     {...other}
   />
 );
-
-OptionsMenuWrapper.propTypes = {
-  children: PropTypes.node.isRequired,
-  className: PropTypes.string,
-  align: PropTypes.oneOf(['left', 'right'])
-};
 
 export default OptionsMenuWrapper;